requirements.txt
smartphone_connector==0.0.107
python-socketio==4.6.0
websocket-client==0.57.0
urllib3==1.25.10
python-engineio==3.13.2
gunicorn==20.0.4
Flask==1.1.2
requests==2.25.1
python-dotenv==0.15.0